move image and calc functions to its corresponding modules
